requirements.txt
inference_sdk
ultralytics
roboflow
hume 
opencv-python 
matplotlib 
pillow
cloudinary
streamlit
streamlit_carousel